Assignment/Day_2/RAG_System_for_Product_Manuals/requirements.txt
flask
groq
pymupdf
faiss-cpu
numpy
sentence-transformers
transformers
torch
rank-bm25
